Rework Kerberos support to not use delegated credentials and support multiple realms.  Still needs more testing and I want to revise the online help.
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@8525 7a7537e8-13f0-0310-91df-b6672ffda945

+++ scheduler/dirsvc.c
@@ -1968,7 +1968,7 @@ cupsdUpdateDNSSDName(void)
     if (btmm && CFGetTypeID(btmm) == CFDictionaryGetTypeID())
     {
       cupsdLogMessage(CUPSD_LOG_DEBUG, "%d Back to My Mac aliases to add.",
-		      CFDictionaryGetCount(btmm));
+		      (int)CFDictionaryGetCount(btmm));
       CFDictionaryApplyFunction(btmm, dnssdAddAlias, NULL);
     }
     else if (btmm)
